fix: move logic outof redundant folders, rename project name and update package versions
