+++ src/c-sharp/Archlens.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34202.233
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Archlens", "Archlens.csproj", "{461DA489-94B8-41E6-8BCB-A3572CBDA98B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|x64.Build.0 = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|x86.Build.0 = Debug|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|x64.ActiveCfg = Release|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|x64.Build.0 = Release|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|x86.ActiveCfg = Release|Any CPU
+		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {04B14143-4302-4414-B85E-CFD4DA3CD9DD}
+	EndGlobalSection
+EndGlobal
