+++ src/c-sharp/Domain/Interfaces/IDependencyParser.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Archlens.Domain.Interfaces;
+
+public interface IDependencyParser
+{
+    Task<IReadOnlyList<string>> ParseModuleDependencies(string path, CancellationToken ct = default);
+    Task<IReadOnlyList<string>> ParseFileDependencies(string path, CancellationToken ct = default);
+}
