+++ src/c-sharp/Domain/Models/Enums/Baseline.cs
@@ -1,0 +1,7 @@
+namespace Archlens.Domain.Models.Enums;
+
+public enum Baseline 
+{
+    Git,
+    Local
+}
