+++ src/c-sharp/Domain/Models/Enums/RenderFormat.cs
@@ -1,0 +1,6 @@
+namespace Archlens.Domain.Models.Enums;
+
+public enum RenderFormat 
+{
+    Json
+}
