--- src/c-sharp/SyntaxTreeManualTraversal/Domain/Interfaces/IRenderer.cs
@@ -1,11 +1,0 @@
-using System.Threading;
-using System.Threading.Tasks;
-using SyntaxTreeManualTraversal.Domain.Models;
-using SyntaxTreeManualTraversal.Domain.Models.Records;
-
-namespace SyntaxTreeManualTraversal.Domain.Interfaces;
-
-public interface IRenderer
-{
-    public Task<string> RenderGraphAsync(DependencyGraph graph, Options options, CancellationToken ct = default);
-}
