--- src/c-sharp/SyntaxTreeManualTraversal/Domain/Models/Enums/Baseline.cs
@@ -1,7 +1,0 @@
-namespace SyntaxTreeManualTraversal.Domain.Models.Enums;
-
-public enum Baseline 
-{
-    Git,
-    Local
-}
