--- src/c-sharp/SyntaxTreeManualTraversal/Domain/Models/Enums/Language.cs
@@ -1,6 +1,0 @@
-namespace SyntaxTreeManualTraversal.Domain.Models.Enums;
-
-public enum Language
-{
-    CSharp
-}
