--- src/c-sharp/SyntaxTreeManualTraversal/Infra/CsharpDependencyParser.cs
@@ -1,19 +1,0 @@
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using SyntaxTreeManualTraversal.Domain.Interfaces;
-
-namespace SyntaxTreeManualTraversal.Infra;
-
-class CsharpDependencyParser : IDependencyParser
-{
-    public Task<IReadOnlyList<string>> ParseFileDependencies(string path, CancellationToken ct = default)
-    {
-        throw new System.NotImplementedException();
-    }
-
-    public Task<IReadOnlyList<string>> ParseModuleDependencies(string path, CancellationToken ct = default)
-    {
-        throw new System.NotImplementedException();
-    }
-}
