--- src/c-sharp/SyntaxTreeManualTraversal/Infra/JsonRenderer.cs
@@ -1,15 +1,0 @@
-using System.Threading;
-using System.Threading.Tasks;
-using SyntaxTreeManualTraversal.Domain.Interfaces;
-using SyntaxTreeManualTraversal.Domain.Models;
-using SyntaxTreeManualTraversal.Domain.Models.Records;
-
-namespace SyntaxTreeManualTraversal.Infra;
-
-public sealed class JsonRenderer : IRenderer
-{
-    public Task<string> RenderGraphAsync(DependencyGraph graph, Options options, CancellationToken ct = default)
-    {
-        throw new System.NotImplementedException();
-    }
-}
