--- src/c-sharp/SyntaxTreeManualTraversal/SyntaxTreeManualTraversal.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34202.233
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SyntaxTreeManualTraversal", "SyntaxTreeManualTraversal.csproj", "{461DA489-94B8-41E6-8BCB-A3572CBDA98B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{461DA489-94B8-41E6-8BCB-A3572CBDA98B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {04B14143-4302-4414-B85E-CFD4DA3CD9DD}
-	EndGlobalSection
-EndGlobal
